Attendance of LegCo Members

Fetch attendance record of LegCo committee meetings

attendance(
  committee_id = NULL,
  meet_id = NULL,
  member_id = NULL,
  attn = "all",
  from = "1900-01-01",
  to = Sys.Date(),
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_attendance(
  committee_id = NULL,
  meet_id = NULL,
  member_id = NULL,
  attn = "all",
  from = "1900-01-01",
  to = Sys.Date(),
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

Arguments

committee_id

the id of a committee, or a vector of ids. If NULL, returns results of all committees. Defaults to NULL.

meet_id

the id of a meeting, or a vector of ids. If NULL, returns all meetings. Useful for matching meeting with records from the Attendance Database. Defaults to NULL.

member_id

the id of a LegCo member, or a vector of ids. If NULL, returns results of all LegCo members. Defaults to NULL.

attn

the attendance. "p" returns members who were present. "a" returns members who were absent. "all" returns all members. Defaults to "all".

from

only fetch results of meetings on or after this date. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to "1900-01-01".

to

only fetch results of meetings on or before this date. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to the current system date.

n

the number of record to fetch. Defaults to 1000.

extra_param

additional query parameters defined in LegCo API. Must begin with "&".

count

logical: Whether to return only the total count of records that matches the parameter(s) instead of the result. Defaults to FALSE.

verbose

logical: Whether to display progress messages when fetching data? Defaults to TRUE.

Details

This function corresponds to the Tattendance data endpoint of the Meeting Attendance Database.

See also

Examples

# \donttest{
# Fetch members who were absent from the Subcommittee on Issues
# Relating to Bazaars meetings on October 13, 2017
x <- attendance(committee_id = 2704, from = "2017-10-13", to = "2017-10-13", attn = "a")
#> Retrieving records...
#> Retrieved 7 record(s). 7 record(s) available in total.
# }