Hansard of LegCo

Fetch metadata and URLs of the hansard files of LegCo council meetings.

hansard(
  hansard_id = NULL,
  lang = "en",
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_hansard(
  hansard_id = NULL,
  lang = "en",
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

Arguments

hansard_id

the id of a hansard file, or a vector of ids. If NULL, returns results of all hansard files. Defaults to NULL.

lang

the language of hansard files to search from. "en" returns the English version. "zh" returns the Traditional Chinese version. Defaults to "en".

from

only fetch results of meetings on or after this date. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to "1900-01-01".

to

only fetch results of meetings on or before this date. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to the current system date.

floor

logical: whether to fetch results from the floor version of the hansard files? The floor version is the first presented version of hansard file in the original language delivered by the speakers in LegCo. If TRUE, the language option is ignored. Defaults to FALSE.

n

the number of record to fetch. Defaults to 1000.

extra_param

additional query parameters defined in LegCo API. Must begin with "&".

count

logical: Whether to return only the total count of records that matches the parameter(s) instead of the result. Defaults to FALSE.

verbose

logical: Whether to display progress messages when fetching data? Defaults to TRUE.

Details

This function corresponds to the Hansard data endpoint of the Hansard Database.

Functions

Functions of the Hansard database:

hansard: Hansard files

legco_section_type: Section code

subjects: Subjects speakers: Speakers in the council, including members, government officials and secretariat staff

rundown: Rundown (Paragraphs in hansard)

questions: Questions raised by members

bills: Bills

motions: Motions

petitions: Petitions

addresses: Addresses made by members or government officials when presenting papers to the Council

statements: Statements made by government officials

voting_results: Results of votes in council meetings

summoning_bells: Instances of summoning bells being rung

See also

Examples

# \donttest{
# Fetch metadata of hansard files of all LegCo meetings
# conducted between Feburary 20 to March 20, 2019
x <- hansard(from = "2019-02-20", to = "2019-03-20")
#> Retrieving records...
#> Retrieved 6 record(s). 6 record(s) available in total.
# }