Meeting Schedule Database

This database contains the information of LegCo committees, namely the name of all LegCo committees established since the fifth term and their corresponding member lists and meeting schedule.

Arguments

member_id
the id of a LegCo member, or a vector of ids. If NULL, returns results of all LegCo members. Defaults to NULL.

term_id
the id of a term, or a vector of ids. If NULL, returns results of all terms. Defaults to NULL.

session_id
the id of a session, or a vector of ids. If NULL, returns result of all sessions. Defaults to NULL.

slot_id
the id of a meeting slot, or a vector of ids. If NULL, returns all meetings. Defaults to NULL.

meet_id
the id of a meeting, or a vector of ids. If NULL, returns all meetings. Useful for matching meeting with records from the Attendance Database. Defaults to NULL.

committee_id
the id of a committee, or a vector of ids. If NULL, returns results of all committees. Defaults to NULL.

code
the code of a committee, or a vector of ids. If NULL, returns all committees. Defaults to NULL.

type
the type of meeting. If "open", returns open meetings. If "closed", returns closed meetings. If "all", returns all meetings. Defaults to "all".

post
the post of a member in the committee, or a vector of posts. "President", "Chairman", "Deputy Chairman", "Member" and "Convenor" return members of the respective post. If NULL, returns members of all posts. Defaults to NULL.

date
only fetch the result in which the specified date falls within. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to NULL.

extra_param
additional query parameters defined in LegCo API. Must begin with "&".

n
the number of record to fetch. Defaults to 1000.

count
logical: Whether to return only the total count of records that matches the parameter(s) instead of the result. Defaults to FALSE.

verbose
logical: Whether to display progress messages when fetching data? Defaults to TRUE.

Functions

Functions of the Meeting Schedule Database:

term: LegCo terms

session: LegCo sessions

committee: LegCo committees

membership: Membership of LegCo committees

member: LegCo members

member_term: Terms served by LegCo members

meeting: Meetings of LegCo committees

meeting_committee: Committees of LegCo meetings

See also