Membership of LegCo Committees

Fetch the members of LegCo committees.

membership(
  member_id = NULL,
  committee_id = NULL,
  term_id = NULL,
  post = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_membership(
  member_id = NULL,
  committee_id = NULL,
  term_id = NULL,
  post = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

Arguments

member_id

the id of a LegCo member, or a vector of ids. If NULL, returns results of all LegCo members. Defaults to NULL.

committee_id

the id of a committee, or a vector of ids. If NULL, returns results of all committees. Defaults to NULL.

term_id

the id of a term, or a vector of ids. If NULL, returns results of all terms. Defaults to NULL.

post

the post of a member in the committee, or a vector of posts. "President", "Chairman", "Deputy Chairman", "Member" and "Convenor" return members of the respective post. If NULL, returns members of all posts. Defaults to NULL.

n

the number of record to fetch. Defaults to 1000.

extra_param

additional query parameters defined in LegCo API. Must begin with "&".

count

logical: Whether to return only the total count of records that matches the parameter(s) instead of the result. Defaults to FALSE.

verbose

logical: Whether to display progress messages when fetching data? Defaults to TRUE.

Details

This function corresponds to the Tmembership data endpoint of the Meeting Schedule Database.

Functions

Functions of the Meeting Schedule Database:

term: LegCo terms

session: LegCo sessions

committee: LegCo committees

membership: Membership of LegCo committees

member: LegCo members

member_term: Terms served by LegCo members

meeting: Meetings of LegCo committees

meeting_committee: Committees of LegCo meetings

See also

Examples

# \donttest{
# Fetch members of the Subcommittee on Issues Relating to Bazaars
x <- membership(committee_id = 2704)
#> Retrieving records...
#> Error: The request did not return any data: Invalid search parameters or rate limit exceeded.
# }