Search Committee

Search committee by full or partial name of committee.

search_committee(search_string, term_id = NULL, exact = TRUE, verbose = TRUE)

legco_search_committee(
  search_string,
  term_id = NULL,
  exact = TRUE,
  verbose = TRUE
)

Arguments

search_string

search string of committee name. Accepts Chinese or English full or partial name. Defaults to NULL.

term_id

the id of a term, or a vector of ids. If NULL, returns results of all terms. Defaults to NULL.

exact

logical: Whether to look for exact match of the search term. Defaults to TRUE.

verbose

logical: Whether to display progress messages when fetching data? Defaults to TRUE.

Details

This is a complementary function to voting_record().

Functions

Complementary Functions:

search_committee: Search LegCo committees

search_member: Search LegCo members

search_voting_record: Search Voting Record in LegCo meetings

search_question: Search full text of question put to the government by LegCo members

Notes

The complementary functions work by calling a number of database functions and combining their output, meaning that a function call usually involves multiple API calls. Use with caution to prevent reaching the API’s rate limit too quickly.

Examples

# \donttest{
# Search all Bills Committees established in the sixth term of LegCo
x <- search_committee(search_string = "bills committee", term_id = 5)
#> Retrieving records...
#> Retrieved 667 record(s). 667 record(s) available in total.
#> 99 record(s) match(es) your parameters.
# }