Search Question in LegCo

Search full text of questions put to the government by LegCo member(s).

search_question(
  speaker_id = NULL,
  member_id = NULL,
  rundown_id = NULL,
  type = "all",
  lang = "en",
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 50,
  verbose = TRUE
)

legco_search_question(
  speaker_id = NULL,
  member_id = NULL,
  rundown_id = NULL,
  type = "all",
  lang = "en",
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 50,
  verbose = TRUE
)

Arguments

speaker_id

the id of a speaker at the Legislative Council, or a vector of ids. If NULL, returns all speakers. Defaults to NULL.

member_id

the id of a LegCo member, or a vector of ids. If NULL, returns results of all LegCo members. Defaults to NULL.

rundown_id

the id of a rundown, or a vector of ids. If NULL, returns results of all rundowns. Defaults to NULL.

type

the type of meeting. If "open", returns open meetings. If "closed", returns closed meetings. If "all", returns all meetings. Defaults to "all".

lang

the language of hansard files to search from. "en" returns the English version. "zh" returns the Traditional Chinese version. Defaults to "en".

from

only fetch results of meetings on or after this date. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to "1900-01-01".

to

only fetch results of meetings on or before this date. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to the current system date.

floor

logical: whether to fetch results from the floor version of the hansard files? The floor version is the first presented version of hansard file in the original language delivered by the speakers in LegCo. If TRUE, the language option is ignored. Defaults to FALSE.

n

the number of record to fetch. Defaults to 50.

verbose

logical: Whether to display progress messages when fetching data? Defaults to TRUE.

Details

This is a complementary function to questions().

Functions

Complementary Functions:

search_committee: Search LegCo committees

search_member: Search LegCo members

search_voting_record: Search Voting Record in LegCo meetings

search_question: Search full text of question put to the government by LegCo members

Notes

The complementary functions work by calling a number of database functions and combining their output, meaning that a function call usually involves multiple API calls. Use with caution to prevent reaching the API’s rate limit too quickly.

Examples

# \donttest{
# Fetch all written questions asked by MemberID 273 from January to April 2019 in Chinese
x <- search_question(member_id = 273, from = "2019-01-01", to ="2019-04-30", type = "written", lang = "zh")
#> Retrieving records...
#> Retrieved 232 record(s). 232 record(s) available in total.
#> Retrieving records...
#> Retrieved 117 record(s). 117 record(s) available in total.
#> Retrieving records...
#> Retrieved 161 record(s). 161 record(s) available in total.
#> 2 record(s) match(es) your parameters.
#> Retrieving records...
#> Retrieved 1 record(s). 1 record(s) available in total.
#> Retrieving records...
#> Retrieved 1 record(s). 1 record(s) available in total.
#> Retrieving records...
#> Retrieved 40 record(s). 40 record(s) available in total.
#> Retrieving records...
#> Retrieved 2 record(s). 2 record(s) available in total.
#> 1 record(s) match(es) your parameters.
# }