Speakers at LegCo

Fetch the basic information of speakers in LegCo council meetings, including LegCo members, government officials and Secretariat staff.

speakers(
  speaker_id = NULL,
  type = "all",
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_speakers(
  speaker_id = NULL,
  type = "all",
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

Arguments

speaker_id

the id of a speaker at the Legislative Council, or a vector of ids. If NULL, returns all speakers. Defaults to NULL.

type

the position of a speaker. "all" returns all speakers. "PO" returns public officers. "LC" returns key appointment holders and staff at LegCo, such as President, Chairman and clerk. "MB" returns LegCo members. Default to "all".

extra_param

additional query parameters defined in LegCo API. Must begin with "&".

count

logical: Whether to return only the total count of records that matches the parameter(s) instead of the result. Defaults to FALSE.

verbose

logical: Whether to display progress messages when fetching data? Defaults to TRUE.

Details

This function corresponds to the Speakers data endpoint of the Hansard Database.

Functions

Functions of the Hansard database:

hansard: Hansard files

legco_section_type: Section code

subjects: Subjects speakers: Speakers in the council, including members, government officials and secretariat staff

rundown: Rundown (Paragraphs in hansard)

questions: Questions raised by members

bills: Bills

motions: Motions

petitions: Petitions

addresses: Addresses made by members or government officials when presenting papers to the Council

statements: Statements made by government officials

voting_results: Results of votes in council meetings

summoning_bells: Instances of summoning bells being rung

See also

Examples

# \donttest{
# Fetch a list of all speakers in LegCo council meetings
x <- speakers()
#> Retrieving records...
#> Retrieved 232 record(s). 232 record(s) available in total.
# Fetch a list of all speakers who are LegCo members
x <- speakers(type = "MB")
#> Retrieving records...
#> Retrieved 182 record(s). 182 record(s) available in total.
# Look up the details of a speaker with the id 6
x <- speakers(speaker_id = 6)
#> Retrieving records...
#> Retrieved 1 record(s). 1 record(s) available in total.
# }