Hansard of LegCo

Fetch metadata and URLs of the hansard files of LegCo council meetings.

hansard(id = NULL, lang = "en", from = "1900-01-01",
  to = Sys.Date(), floor = FALSE, n = 1000, extra_param = NULL,
  count = FALSE, verbose = TRUE)

legco_hansard(id = NULL, lang = "en", from = "1900-01-01",
  to = Sys.Date(), floor = FALSE, n = 1000, extra_param = NULL,
  count = FALSE, verbose = TRUE)

Arguments

id
The id of a hansard file, or a vector of ids. If `NULL`, returns a list of all hansard files. Defaults to `NULL`.

lang
The language of hansard files to search from. `'en'` returns the English version. `'zh'` returns the Traditional Chinese version. Defaults to `'en'`.

from
Only includes hansards of meetings on or after this date. Accepts character values in `'YYYY-MM-DD'` format, and objects of class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a date with `as.Date()`. Defaults to `'1900-01-01'`.

to
Only includes hansards of meetings on or before this date. Accepts character values in `'YYYY-MM-DD'` format, and objects of class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a date with `as.Date()`. Defaults to the current system date.

floor
Whether to fetch the floor version of the hansard files. The floor version is the first presented version of hansard file in the original language delivered by the speakers in LegCo. If `TRUE`, the language option is ignored. Defaults to `FALSE`.

n
The number of hansard files to request. Defaults to `1000`.

extra_param
Additional query parameters defined in LegCo API. Must begin with `'&'`.

count
If `TRUE`, returns only the total count of records that matches the paramter(s) instead of the result. Defaults to `FALSE`.

verbose
Defaults to `TRUE`.